% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{compute_lse}
\alias{compute_lse}
\title{Compute Least Squares Estimates}
\usage{
compute_lse(X, Y)
}
\arguments{
\item{X}{an n x (p-1) matrix of predictors}

\item{Y}{an n-length vector of the response variable (left out node)}
}
\value{
a vector of coefficient estimates obtained from the least squares solution
}
\description{
Computes least squares estimates for regression coefficients using power expansion of features.
}
