% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphData.R
\name{.gibbs_sample}
\alias{.gibbs_sample}
\title{Run Gibbs sampler}
\usage{
.gibbs_sample(theta, bias, n_samples, burn_in, seed = NULL)
}
\description{
Run Gibbs sampler
}
\keyword{Draw}
\keyword{Ising}
\keyword{and}
\keyword{bias}
\keyword{by}
\keyword{defined}
\keyword{from}
\keyword{internal}
\keyword{model}
\keyword{samples}
\keyword{the}
\keyword{theta}
