% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphData.R
\name{.gibbs_sweep}
\alias{.gibbs_sweep}
\title{Single Gibbs sweep over all nodes (±1 encoding)}
\usage{
.gibbs_sweep(x, theta, bias)
}
\description{
Single Gibbs sweep over all nodes (±1 encoding)
}
\keyword{For}
\keyword{X_i}
\keyword{all}
\keyword{and}
\keyword{bias}
\keyword{current}
\keyword{each}
\keyword{given}
\keyword{i,}
\keyword{internal}
\keyword{node}
\keyword{nodes}
\keyword{of}
\keyword{other}
\keyword{parameters}
\keyword{sample}
\keyword{state}
\keyword{the}
\keyword{theta}
