% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphData.R
\name{generate_ising_to_multinomial}
\alias{generate_ising_to_multinomial}
\title{Generate a multinomial distribution from a specified Theta matrix}
\usage{
generate_ising_to_multinomial(Theta, bias = NULL)
}
\arguments{
\item{bias}{Optional bias vector (length p). If NULL, assumed to be zero.}

\item{theta}{Coupling matrix (p x p) defining the Ising model.}
}
\value{
A data frame with two columns: 'configuration' (binary string) and 'probability' (corresponding probability).
The return follows the ordering of the Walsh/Hadamard transform to conform with the BEStat framework
}
\description{
Generate a multinomial distribution from a specified Theta matrix
}
