% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphData.R
\name{ising_summary}
\alias{ising_summary}
\title{Print a summary of an ising_generate result}
\usage{
ising_summary(result)
}
\arguments{
\item{result}{List returned by \code{ising_generate}.}
}
\description{
Print a summary of an ising_generate result
}
