% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphData.R
\docType{package}
\name{isinggraph-package}
\alias{BEstatGraphTools}
\alias{BEstatGraphTools-package}
\alias{isinggraph-package}
\title{Generate Binary Graph Data with CI Constraints (Ising Model, ±1 encoding)}
\description{
Enforces CI conditions X_i ⊥ X_j | X_rest iff (i,j) is not an edge.
Uses an Ising MRF: P(X) ∝ exp(X^T Θ X / 2 + b^T X), X ∈ {-1, +1}^n.
Bias terms b_i are calibrated so that P(X_i = +1) ≈ target_probs\link{i}.
}
\author{
\strong{Maintainer}: Evan Schwartz \email{eschw@unc.edu}

}
