% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{J_i}
\alias{J_i}
\title{Select the indices for a RV A_i of the vector A to capture the indices of all interactions that include A_i
We only use the dimension of data}
\usage{
J_i(j, p)
}
\arguments{
\item{p}{the number of nodes in the graph or RVs in a BELIEF regression}

\item{i}{the index of the node or RV for which we want to capture interactions}
}
\value{
a vector of indices for the interactions that include A_i
}
\description{
Select the indices for a RV A_i of the vector A to capture the indices of all interactions that include A_i
We only use the dimension of data
}
