% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_detected_graph}
\alias{plot_detected_graph}
\title{Function to plot detected conditional dependence graph}
\usage{
plot_detected_graph(
  avg_results,
  threshold = 0.5,
  main = "Detected Graph at Threshold"
)
}
\arguments{
\item{avg_results}{Data frame with columns node_i, node_j, reject_at_alpha}

\item{threshold}{Numeric. Threshold for reject_at_alpha to consider an edge detected. Default is 0.5. Usually should set to 1-alpha for a specific alpha level.}

\item{main}{Title for the plot. Default is "Detected Graph at Threshold".}
}
\value{
Invisibly returns the igraph object of the detected graph.
}
\description{
Function to plot detected conditional dependence graph
}
