% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_theta_histograms}
\alias{plot_theta_histograms}
\title{Plot histograms of theta values (e.g. level of CI) for specified edges across simulations}
\usage{
plot_theta_histograms(theta_list, edge_list, main_prefix = "Theta Hist")
}
\arguments{
\item{theta_list}{List of theta matrices (p x p) from multiple simulations}

\item{edge_list}{Two-column matrix of node pairs (i, j) for which to plot histograms of theta values, should typically be the ising edges input}

\item{main_prefix}{Prefix for the main title of each histogram, default is "Theta Hist"}
}
\value{
Invisibly returns NULL after plotting histograms
}
\description{
Plot histograms of theta values (e.g. level of CI) for specified edges across simulations
}
