% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{power_expand}
\alias{power_expand}
\title{Functions for following the BEstat framework}
\usage{
power_expand(X)
}
\arguments{
\item{X}{a matrix or vector of binary data, in our case nodes in a graph}
}
\value{
a vector of p probabilities for P(A_i = 1) for i in 1,...,p
}
\description{
Power operator function
}
