% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{wilks_LRT_permutation_test}
\alias{wilks_LRT_permutation_test}
\title{Run an LRT permutation test for binary data}
\usage{
wilks_LRT_permutation_test(A, n_permutations = 1000, seed = NULL)
}
\arguments{
\item{A}{an n x p binary data matrix}

\item{n_permutations}{the number of permutations to perform for estimating the null distribution of the test statistic (default is 1000)}

\item{seed}{an optional random seed for reproducibility of the permutations}
}
\value{
a matrix of p-values for the likelihood ratio test comparing the full model (with all predictors) to the null model (without each predictor) for each node. The diagonal entries are NA since we do not test nodes against themselves and bottom of matrix is NA since we only compute upper-triangular entries to avoid redundant tests.
The matrix is symmetric with NA on the diagonal, with p-values for the tests in the off-diagonal entries to allow for multiple testing control on repeated tests
}
\description{
Run an LRT permutation test for binary data
}
