% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{Wilks_LRT_Test}
\alias{Wilks_LRT_Test}
\title{Run the Wilks Likelihood Ratio Test for Binary Data}
\usage{
Wilks_LRT_Test(A, node_j)
}
\arguments{
\item{A}{an n x p binary data matrix}

\item{node_j}{the index of node we are regressing on}
}
\value{
a list containing the test statistic and p-value for the Wilks LRT
}
\description{
Run the Wilks Likelihood Ratio Test for Binary Data
}
