% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphData.R
\name{.build_theta}
\alias{.build_theta}
\title{Build the coupling matrix from an edge list}
\usage{
.build_theta(n_nodes, edges, weak_edges = NULL, weights = NULL)
}
\description{
Build the coupling matrix from an edge list
}
\keyword{internal}
