% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{wilks_LRT_test}
\alias{wilks_LRT_test}
\title{Run the Wilks Likelihood Ratio Test for Binary Data}
\usage{
wilks_LRT_test(A)
}
\arguments{
\item{A}{an n x p binary data matrix}

\item{node_j}{the index of node we are regressing on}
}
\value{
a matrix of p-values for the likelihood ratio test comparing the full model (with all predictors) to the null model (without each predictor) for node_j. The diagonal entry corresponding to node_j is NA since we do not test the node against itself.
The matrix should be symmetric and have NA on the diagonal, with p-values for the tests in the off-diagonal entries. This hopefully should prevent multiple testing control on repeated tests
}
\description{
Run the Wilks Likelihood Ratio Test for Binary Data
}
