% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data_tools.R
\name{wilks_LRT_test_e_val}
\alias{wilks_LRT_test_e_val}
\title{Construct split likelihood-ratio e-values for binary data}
\usage{
wilks_LRT_test_e_val(A)
}
\arguments{
\item{A}{an n x p binary data matrix}
}
\value{
an upper-triangular matrix of e-values comparing the full model
(with all predictors) to the null model (without each predictor) for each
node. The diagonal and lower-triangular entries are \code{NA}.
}
\description{
Construct split likelihood-ratio e-values for binary data
}
